simplify further

+++ index.js
@@ -1,6 +1,5 @@
 'use strict';
 
-const PQueue = require('p-queue');
 const tilebelt = require('tilebelt');
 const mkdirp = require('mkdirp');
 const hash = require('shorthash').unique;
@@ -19,18 +18,18 @@ const defaultOptions = {
     cacheDir: path.join(__dirname, '/tile-cache')
 };
 
-function runStats(url, processTile, options) {
+async function runStats(url, processTile, options) {
 
     options = Object.assign(Object.create(defaultOptions), options || {});
 
     const cachePath = path.join(options.cacheDir, hash(url));
     mkdirp(cachePath);
 
-    const q = new PQueue({concurrency: 4});
-
     const handleTileLoad = (tile) => {
         if (tile) processTile(tile);
     };
+
+    const tilePromises = [];
 
     for (let z = options.minZoom; z <= options.maxZoom; z++) {
         const p = tilebelt.pointToTileFraction(options.center[0], options.center[1], z);
@@ -43,45 +42,42 @@ function runStats(url, processTile, options) {
 
         for (let x = minX; x <= maxX; x++) {
             for (let y = minY; y <= maxY; y++) {
-                q.add(loadTile(z, x, y, url, cachePath)).then(handleTileLoad);
+                tilePromises.push(loadTile(z, x, y, url, cachePath).then(handleTileLoad));
             }
         }
     }
 
-    return q.onIdle().then(() => {
-        process.stderr.write('\n');
-    });
+    await Promise.all(tilePromises);
+
+    process.stderr.write('\n');
 }
 
-function loadTile(z, x, y, urlTemplate, cachePath) {
-    return async () => {
-        const tilePath = path.join(cachePath, z + '-' + x + '-' + y);
-        try {
-            const data = fs.readFileSync(tilePath);
-            process.stderr.write('.');
-            return {z, x, y, data};
+async function loadTile(z, x, y, urlTemplate, cachePath) {
+    const tilePath = path.join(cachePath, z + '-' + x + '-' + y);
 
-        } catch (e) {
-            const url = urlTemplate
-                .replace('{x}', x)
-                .replace('{y}', y)
-                .replace('{z}', z);
+    if (fs.existsSync(tilePath)) {
+        const data = fs.readFileSync(tilePath);
+        process.stderr.write('.');
+        return {z, x, y, data};
+    }
 
-            const response = await fetch(url);
+    const url = urlTemplate
+        .replace('{x}', x)
+        .replace('{y}', y)
+        .replace('{z}', z);
 
-            if (response.status === 200) {
-                const data = await response.buffer();
-                process.stderr.write('+');
-                fs.writeFileSync(tilePath, data);
-                return {z, x, y, data};
+    const response = await fetch(url);
 
-            } else if (response.status === 404) {
-                process.stderr.write('_');
+    if (response.status === 200) {
+        const data = await response.buffer();
+        process.stderr.write('+');
+        fs.writeFileSync(tilePath, data);
+        return {z, x, y, data};
 
-            } else {
-                throw new Error(response.status + ' ' + url);
-            }
-        }
-        return null;
-    };
+    } else if (response.status === 404) {
+        process.stderr.write('_');
+
+    } else {
+        throw new Error(response.status + ' ' + url);
+    }
 }
